# internal/entity/user.go
package entity

import (
	"time"
)

type Role string

const (
	RoleUser  Role = "user"
	RoleAdmin Role = "admin"
)

// User -.
type User struct {
	ID        int       `json:"id"`
	FullName  string    `json:"fullname"`
	Email     string    `json:"email"`
	Phone     string    `json:"phone"`
	Password  string    `json:"-"`
	Token     string    `json:"token"`
	Role      Role      `json:"role"`
	CreatedAt time.Time `json:"created_at"`
	UpdatedAt time.Time `json:"updated_at"`
}

type UserRegister struct {
	FullName string `json:"fullname"`
	Email    string `json:"email"`
	Password string `json:"password_hash"`
}

type UserUpdate struct {
	ID       int    `json:"id"`
	FullName string `json:"fullname"`
	Email    string `json:"email"`
	Password string `json:"password_hash"`
	Phone    string `json:"phone"`
}

type GetPasswordHash struct {
	PasswordHash string `json:"-"`
	ID           int    `json:"id"`
}